Refactor: Move DukanApiImpl to api package
