+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/api/DukanApiImpl.kt
@@ -1,7 +1,8 @@
-package net.thechance.mena.dukan.presentation.navigation
+package net.thechance.mena.dukan.presentation.api
 
 import androidx.compose.runtime.Composable
 import net.thechance.mena.dukan.api.DukanApi
+import net.thechance.mena.dukan.presentation.navigation.DukanNavHost
 
 class DukanApiImpl: DukanApi {
     @Composable
