+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/di/DukanApiModule.kt
@@ -1,7 +1,7 @@
 package net.thechance.mena.dukan.presentation.di
 
 import net.thechance.mena.dukan.api.DukanApi
-import net.thechance.mena.dukan.presentation.navigation.DukanApiImpl
+import net.thechance.mena.dukan.presentation.api.DukanApiImpl
 import org.koin.core.module.dsl.factoryOf
 import org.koin.dsl.bind
 import org.koin.dsl.module
